% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.build.program}
\alias{cl.build.program}
\title{Build OpenCL program}
\usage{
cl.build.program(program, options = NULL)
}
\arguments{
\item{program}{An opencl_program object}

\item{options}{OpenCL build options as a string}
}
\description{
Compiles the code in an OpenCL program object.
}
