% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.create.buffer}
\alias{cl.create.buffer}
\title{Allocate OpenCL buffer}
\usage{
cl.create.buffer(context, flags, size = NULL, vector = NULL)
}
\arguments{
\item{context}{An opencl_context object.}

\item{flags}{Character vector of OpenCL buffer flags. Refer to \href{https://www.khronos.org/registry/OpenCL/sdk/2.0/docs/man/xhtml/clCreateBuffer.html}{The OpenCL documentation}.}

\item{size}{The size in bytes of the vector. Mandatory if vector is NULL.}

\item{vector}{The data to go into an OpenCL buffer. Mandatory if size is NULL.}
}
\value{
An opencl_buffer object
}
\description{
Allocates an OpenCL buffer with the given context.
}
\details{
Calls clCreateBuffer with the given context, flags and buffer.
The data is optional, if vector is left null an empty buffer wil be allocated. It can be 
filled with data at a later point with a call to cl.enqueue.write.buffer.
The size must be the size in bytes of the OpenCL buffer to be allocated. Note that 
object.size includes R overhead. Integer and float32 vectors take up 4 bytes
per element, double vectors (the R default) take up 8 bytes per element. 
Note that 64-bit float performance is crippled on most GPUs, so it is a good idea
to cast a double vector to a float32 vector before use with OpenCL.
}
