% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.create.context}
\alias{cl.create.context}
\title{Create OpenCL context}
\usage{
cl.create.context(devices)
}
\arguments{
\item{devices}{One or a list of opencl_device objects.}
}
\value{
An opencl_context object.
}
\description{
Allocates an OpenCL context with one or more devices.
}
\details{
A finalizer is not registered for contexts, thus they must be freed manually. 
Contexts must be freed only after all OpenCL objects created with the context are
freed.
}
