% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.create.kernel}
\alias{cl.create.kernel}
\title{Get OpenCL kernel}
\usage{
cl.create.kernel(program, name)
}
\arguments{
\item{program}{An opencl_program object}

\item{name}{The name of the kernel to get.}
}
\description{
Extracts an OpenCL kernel from a program
}
\details{
Finds the specified kernel in the given program. The program must have been built.
}
