% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.create.program}
\alias{cl.create.program}
\title{Create OpenCL program}
\usage{
cl.create.program(context, source_code, build = F)
}
\arguments{
\item{context}{An opencl_context object}

\item{source_code}{The source code of the OpenCL kernel}

\item{build}{Automatically build the program}
}
\value{
An opencl_program object
}
\description{
Creates an OpenCL program from source code.
}
