% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.create.queue}
\alias{cl.create.queue}
\title{Create an OpenCL command queue}
\usage{
cl.create.queue(context, device, in_order = T, profiling = F,
  on_device = F)
}
\arguments{
\item{context}{An opencl_context object}

\item{device}{An opencl_device object}

\item{in_order}{Specifies whether the queue is an in-order or out-of-order queue. 
Not currently used, included for future use.}

\item{profiling}{Enable profiling of commands. Not currently used, included for future use.}

\item{on_device}{Create a device-side queue. Not currently used, included for future use.}
}
\value{
An opencl_command_queue object.
}
\description{
Creates an OpenCL command queue for the given context and device.
}
