% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.device.getattr}
\alias{cl.device.getattr}
\title{Get OpenCL device info}
\usage{
cl.device.getattr(device, attr)
}
\arguments{
\item{device}{An opencl_device object}

\item{attr}{An OpenCL device attribute as a string}
}
\value{
Requested information about the device
}
\description{
Gets information from clGetDeviceInfo call
}
