% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.devices.count}
\alias{cl.devices.count}
\title{Device Count}
\usage{
cl.devices.count(platform)
}
\arguments{
\item{platform}{An opencl_platform object}
}
\value{
The number of devices available
}
\description{
Counts the number of available devices belonging to a platform
}
