% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.enqueue.read.buffer}
\alias{cl.enqueue.read.buffer}
\title{Read data from a buffer}
\usage{
cl.enqueue.read.buffer(queue, buffer, destination, offset = 0L,
  blocking = T, waitlist = NULL)
}
\arguments{
\item{queue}{An opencl_command_queue object. The queue the operation will be put on.}

\item{buffer}{An opencl_buffer object. The buffer to read from.}

\item{destination}{The destination vector to write data into. Must be as long in bytes as the data being read.}

\item{offset}{Offset in bytes from the beginning of the buffer to begin reading from.}

\item{blocking}{Whether to wait for the write to finish or return immediately. 
Not currently used, included for future use.}

\item{waitlist}{OpenCL events that must complete before this operation can start.
Not currently used, included for future use.}
}
\description{
Enqueues an operation to read data from an OpenCL buffer to an R vector.
}
