% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.enqueue.write.buffer}
\alias{cl.enqueue.write.buffer}
\title{Write data to a buffer}
\usage{
cl.enqueue.write.buffer(queue, buffer, vector, blocking = T,
  offset = 0L, waitlist = NULL)
}
\arguments{
\item{queue}{An opencl_command_queue object. The queue the operation will be put on.}

\item{buffer}{An opencl_buffer object. The buffer to write data to.}

\item{vector}{The data to write. The length in bytes of the vector must not be longer than the buffer.}

\item{blocking}{Whether to wait for the write to finish or return immediately. 
Not currently used, included for future use.}

\item{offset}{Offset in bytes from the beginning of the supplied vector.}

\item{waitlist}{OpenCL events that must complete before this operation can start.
Not currently used, included for future use.}
}
\description{
Enqueues an operation to write data to an OpenCL buffer.
}
