% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.get.devices}
\alias{cl.get.devices}
\title{Get OpenCL devices}
\usage{
cl.get.devices(platform)
}
\arguments{
\item{platform}{An opencl_platform object}
}
\value{
A list of opencl_device objects
}
\description{
Gets all OpenCL devices available under a given platform
}
