% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.kernel.getattr}
\alias{cl.kernel.getattr}
\title{Get OpenCL kernel info}
\usage{
cl.kernel.getattr(kernel, what)
}
\arguments{
\item{what}{An OpenCL kernel attribute as a string}

\item{device}{An opencl_kernel object}
}
\value{
Requested information about the kernel
}
\description{
Gets information from clGetKernelInfo call
}
