% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.load.binary}
\alias{cl.load.binary}
\title{Load OpenCL binary}
\usage{
cl.load.binary(path, context, devs)
}
\arguments{
\item{path}{File path of the binary}

\item{context}{An opencl_context object}

\item{devs}{A list of devices to load the binary into}
}
\value{
An opencl_program object
}
\description{
Creates a program from a pre-compiled OpenCL binary
}
\details{
Loads an OpenCL binary. Can only load a binary into the same device as
the binary was built with. So a compiled CPU kernel cannot be loaded into a GPU context.
}
