% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.platform.getattr}
\alias{cl.platform.getattr}
\title{Get OpenCL platform info}
\usage{
cl.platform.getattr(device, attr)
}
\arguments{
\item{device}{An opencl_platform object}

\item{attr}{An OpenCL platform attribute as a string}
}
\value{
Requested information about the platform
}
\description{
Gets information from clGetPlatformInfo call
}
