% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.release.buffer}
\alias{cl.release.buffer}
\title{Release OpenCL buffer}
\usage{
cl.release.buffer(buffer)
}
\arguments{
\item{buffer}{An opencl_buffer object}
}
\description{
Frees the resources held by the OpenCL buffer object.
}
