% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.release.context}
\alias{cl.release.context}
\title{Release OpenCL context}
\usage{
cl.release.context(context)
}
\arguments{
\item{context}{An opencl_context object}
}
\description{
Releases resources held by the OpenCL context.
}
\details{
Context must be released last, after all other objects are released (or garbage collected)
}
