% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.release.kernel}
\alias{cl.release.kernel}
\title{Release OpenCL kernel}
\usage{
cl.release.kernel(kernel)
}
\arguments{
\item{kernel}{An opencl_kernel object}
}
\description{
Frees the resources held by the OpenCL kernel object.
}
