% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.release.program}
\alias{cl.release.program}
\title{Release OpenCL program}
\usage{
cl.release.program(program)
}
\arguments{
\item{program}{An opencl_program object}
}
\description{
Frees the resources held by the OpenCL program object.
}
