% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcl.r
\name{cl.release.queue}
\alias{cl.release.queue}
\title{Release OpenCL queue}
\usage{
cl.release.queue(queue)
}
\arguments{
\item{queue}{An opencl_queue object}
}
\description{
Frees the resources held by the OpenCL queue object.
}
